Moves docstring into string.
